requirements.txt
scikit-learn
langchain_community
numpy
faiss-cpu
PyPDF2
PyCryptodome